# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: AnyLive API
  version: 0.1.0
  description: |
    AnyLive overseas-first live streaming control plane.
    P0 surface: system health + meta. Auth/rooms freeze in v0.1/v0.2.
servers:
  - url: http://localhost:8088
    description: Local API
...
paths:
...
  /api/v1/wallet:
    get:
      tags: [wallet]
      operationId: getWallet
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: Balance
          content:
            application/json:
              schema:
                type: object
                required: [balance]
                properties:
                  balance: {type: integer, format: int64}
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/wallet/ledger:
    get:
      tags: [wallet]
      operationId: getWalletLedger
      summary: List wallet ledger entries for current user
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: Ledger entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerList"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/wallet/topups:
    post:
      tags: [wallet]
      operationId: topupWallet
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount: {type: integer, format: int64}
      responses:
        "200": {description: Updated balance}
...
  /api/v1/rooms/{roomId}/gifts:
    post:
      tags: [gifts]
      operationId: sendGift
      security: [{bearerAuth: []}]
      parameters:
        - $ref: "#/components/parameters/RoomId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [gift_id, receiver_id, client_request_id]
              properties:
                gift_id: {type: string, format: uuid}
                receiver_id: {type: string, format: uuid}
                count: {type: integer, default: 1}
                client_request_id: {type: string}
      responses:
        "201": {description: Created}
        "200": {description: Idempotent replay}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RoomId:
      name: roomId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      required: [code, message, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        retryable:
          type: boolean
...
    LedgerEntry:
      type: object
      required: [id, amount, balance_after, entry_type, reference, created_at]
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: integer
          format: int64
        balance_after:
          type: integer
          format: int64
        entry_type:
          type: string
          enum: [topup, gift_debit, gift_credit, adjustment]
        reference:
          type: string
        created_at:
          type: string
          format: date-time
    LedgerList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/LedgerEntry"
...